Rewrite xur service and refactoring guild features
