--- Neuromatrix/Services/RateLimitService.cs
@@ -1,77 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.Concurrent;
-
-using Discord;
-using Discord.Commands;
-
-namespace Neuromatrix.Services
-{
-    public class RateLimitService
-    {
-        #region Private fields
-        private ConcurrentDictionary<string, RateLimitInfo> _ratelimits;
-        #endregion
-
-        public RateLimitService()
-        {
-            _ratelimits = new ConcurrentDictionary<string, RateLimitInfo>();
-        }
-
-        public RateLimitInfo GetOrAdd(string rule, Func<string, RateLimitInfo> createFunc)
-            => _ratelimits.GetOrAdd(rule, createFunc);
-
-        /// <summary>
-        /// Get a RateLimit rule for global ratelimits.
-        /// </summary>
-        public string GetRule(CommandInfo command)
-            => $"{command.Module.Name}#{command.Name}";
-        /// <summary>
-        /// Get a RateLimit rule for user ratelimits.
-        /// </summary>
-        public string GetActorRule(IUser actor, CommandInfo command)
-            => $"actor:{actor.Id}/{GetRule(command)}";
-        /// <summary>
-        /// Get a RateLimit rue for channel ratelimits.
-        /// </summary>
-        public string GetChanelRule(IChannel channel, CommandInfo command)
-            => $"channel:{channel.Id}/{GetRule(command)}";
-    }
-
-    public class RateLimitInfo
-    {
-        #region Public fields
-        public uint Capacity { get; private set; }
-        public TimeSpan DrainRate { get; private set; }
-        public HashSet<DateTimeOffset> Entries { get; private set; }
-        #endregion
-
-        public RateLimitInfo(uint capacity, TimeSpan drainRate)
-        {
-            Capacity = capacity;
-            DrainRate = drainRate;
-            Entries = new HashSet<DateTimeOffset>();
-        }
-
-        public bool CanEnter()
-        {
-            int current = 0;
-            var now = DateTimeOffset.Now;
-            foreach (var entry in Entries)
-            {
-                if (now - entry > DrainRate)
-                    Entries.Remove(entry);
-                else
-                    current++;
-            }
-            if (current >= Capacity)
-                return false;
-            return true;
-        }
-
-        public void Increment()
-        {
-            Entries.Add(DateTimeOffset.Now);
-        }
-    }
-}
