+++ Neuromatrix/Services/ReminderService.cs
@@ -29,9 +29,13 @@ namespace Neuromatrix.Services
         {
             // If signal time equal Friday 20:00 we will send message Xur is arrived in game.
             if (e.SignalTime.DayOfWeek == DayOfWeek.Friday && e.SignalTime.Hour == 20 && e.SignalTime.Minute == 00 && e.SignalTime.Second < 10)
+                //Test version
+                //if (e.SignalTime.DayOfWeek == DayOfWeek.Tuesday && e.SignalTime.Hour == 15 && e.SignalTime.Minute == 46 && e.SignalTime.Second < 10)
                 await XurArrived();
             // If signal time equal Tuesday 20:00 we will send message Xur is leave game.
             if (e.SignalTime.DayOfWeek == DayOfWeek.Tuesday && e.SignalTime.Hour == 20 && e.SignalTime.Minute == 00 && e.SignalTime.Second < 10)
+                //Test version
+                //if (e.SignalTime.DayOfWeek == DayOfWeek.Tuesday && e.SignalTime.Hour == 15 && e.SignalTime.Minute == 41 && e.SignalTime.Second < 10)
                 await XurLeave();
         }
 
@@ -44,10 +48,11 @@ namespace Neuromatrix.Services
                 .WithThumbnailUrl("http://159.69.21.188/Icon/xur_emblem.png")
                 .WithDescription("Нажмите на заголовок сообщения чтобы узнать точное местоположение посланника Зура.")
                 .WithUrl("https://whereisxur.com/")
-                .WithFooter("Напоминаю! Зур покинет солнечную систему во вторник в 20:00 по МСК.");
+                .WithFooter("Напоминаю! Зур покинет солнечную систему во вторник в 20:00 по МСК.")
+                .WithCurrentTimestamp();
             #endregion
 
-            var guilds = Database.GetAllGuilds();
+            var guilds = await Database.GetAllGuildsAsync();
 
             foreach (var guild in guilds)
             {
@@ -75,10 +80,11 @@ namespace Neuromatrix.Services
                .WithTitle("Внимание! Зур покинул солнечную систему.")
                .WithThumbnailUrl("http://159.69.21.188/Icon/xur_emblem.png")
                .WithDescription("Он просто испарился! Как только он придёт я сообщу.")
-               .WithFooter("Напоминаю! В следующий раз Зур прибудет в пятницу в 20:00 по МСК.");
+               .WithFooter("Напоминаю! В следующий раз Зур прибудет в пятницу в 20:00 по МСК.")
+               .WithCurrentTimestamp();
             #endregion
 
-            var guilds = Database.GetAllGuilds();
+            var guilds = await Database.GetAllGuildsAsync();
 
             foreach (var guild in guilds)
             {
